% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFSreconstruction.R
\name{AFSreconstruction}
\alias{AFSreconstruction}
\title{Advancing front surface reconstruction}
\usage{
AFSreconstruction(points)
}
\arguments{
\item{points}{numeric matrix which stores the points, one point per row}
}
\value{
A triangular mesh, of class \code{mesh3d} (ready for plotting
  with \strong{rgl}).
}
\description{
Reconstruction of a surface from a cloud of 3D points.
}
\examples{
library(RCGAL)
data(bunny, package = "onion")
mesh <- AFSreconstruction(bunny)
library(rgl)
shade3d(mesh, color = "firebrick")
}
