% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-CostaSurface.R
\docType{data}
\name{CostaSurface}
\alias{CostaSurface}
\title{A points cloud of the Costa surface}
\format{
A matrix with 2542 rows and 3 columns.
}
\usage{
CostaSurface
}
\description{
A matrix containing some points forming the Costa surface.
}
\keyword{datasets}
