% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cyclide.R
\docType{data}
\name{cyclide}
\alias{cyclide}
\title{A points cloud of a cyclide}
\format{
A matrix with 3708 rows and 3 columns.
}
\usage{
cyclide
}
\description{
A matrix containing some points forming a cyclide.
}
\keyword{datasets}
