% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delaunay.R
\name{delaunay}
\alias{delaunay}
\title{Delaunay tessellation}
\usage{
delaunay(points)
}
\arguments{
\item{points}{numeric matrix which stores the points, one point per row}
}
\value{
The Delaunay tessellation.
}
\description{
Delaunay tessellation of a set of 2D or 3D points.
}
\examples{
library(RCGAL)
}
