% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normals.R
\name{getSomeNormals}
\alias{getSomeNormals}
\title{Normals for a points could}
\usage{
getSomeNormals(points, nbNeighbors)
}
\arguments{
\item{points}{matrix of points, one point per row}

\item{nbNeighbors}{integer, number of neighbors used to compute the normals}
}
\value{
A matrix of the same size as the \code{points} matrix, giving one
  unit normal for each point.
}
\description{
Compute some normals for a 3D points cloud.
}
