% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-HopfTorus.R
\docType{data}
\name{HopfTorus}
\alias{HopfTorus}
\title{A points cloud of a Hopf torus}
\format{
A matrix with 4142 rows and 3 columns.
}
\usage{
HopfTorus
}
\description{
A matrix containing some points forming (the stereographic
  projection of) a Hopf torus.
}
\keyword{datasets}
