% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ICN5D_eight.R
\docType{data}
\name{ICN5D_eight}
\alias{ICN5D_eight}
\title{A points cloud of ICN5D's eight-like surface}
\format{
A matrix with 11736 rows and 3 columns.
}
\usage{
ICN5D_eight
}
\description{
A matrix containing some points forming ICN5D's eight-like
  surface. ICN5D is the pseudonym of a member of the
  \href{http://hi.gher.space/forum/}{Hi.gher. Space forum}.
}
\keyword{datasets}
