% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{Mesh}
\alias{Mesh}
\title{xxx}
\usage{
Mesh(vertices, faces, normals = TRUE)
}
\arguments{
\item{vertices}{xxx}
\item{faces}{xxx}
\item{normals}{xxx}
}
\value{
xxx
}
\description{
xxx
}
\examples{
library(RCGAL)
# a tetrahedron with ill-oriented faces:
vertices <- rbind(
  c(-1, -1, -1),
  c(1, 1, -1),
  c(1, -1, 1),
  c(-1, 1, 1)
)
faces <- rbind(
  c(1, 2, 3),
  c(3, 4, 2),
  c(4, 2, 1),
  c(4, 3, 1)
)
Mesh(vertices, faces, normals = FALSE)
}
