% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{MeshesIntersection}
\alias{MeshesIntersection}
\title{Meshes intersection}
\usage{
MeshesIntersection(meshes, merge = FALSE, normals = FALSE, exact = FALSE)
}
\arguments{
\item{meshes}{a list of \emph{triangular} meshes, each given as a list with
(at least) two fields: \code{vertices} and \code{faces}}

\item{merge}{Boolean, whether to merge the duplicated vertices of the
input meshes and the output mesh}

\item{normals}{Boolean, whether to return the per-vertex normals of the
output mesh}

\item{exact}{Boolean, whether to use exact calculations; this is slower but
more accurate}
}
\value{
A triangular mesh given as a list with fields \code{vertices},
  \code{faces}, \code{edges} and \code{normals} if \code{normals=TRUE}.
}
\description{
Computes the intersection of the given meshes.
}
\examples{
library(RCGAL)
library(rgl)
}
