% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Orthocircle.R
\docType{data}
\name{Orthocircle}
\alias{Orthocircle}
\title{A points cloud of an orthocircle}
\format{
A matrix with 3996 rows and 3 columns.
}
\usage{
Orthocircle
}
\description{
A matrix containing some points forming an orthocircle.
}
\keyword{datasets}
