% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{plotEdges}
\alias{plotEdges}
\title{Plot some edges}
\usage{
plotEdges(
  vertices,
  edges,
  color = "black",
  lwd = 2,
  edgesAsTubes = TRUE,
  tubesRadius = 0.03,
  verticesAsSpheres = TRUE,
  spheresRadius = 0.05,
  spheresColor = color
)
}
\arguments{
\item{vertices}{a three-columns matrix giving the coordinates of the vertices}

\item{edges}{a two-columns integer matrix giving the edges by pairs of
vertex indices}

\item{color}{a color for the edges}

\item{lwd}{line width, a positive number, ignored if \code{edgesAsTubes=TRUE}}

\item{edgesAsTubes}{Boolean, whether to draw the edges as tubes}

\item{tubesRadius}{the radius of the tubes when \code{edgesAsTubes=TRUE}}

\item{verticesAsSpheres}{Boolean, whether to draw the vertices as spheres}

\item{spheresRadius}{the radius of the spheres when
\code{verticesAsSpheres=TRUE}}

\item{spheresColor}{the color of the spheres when
\code{verticesAsSpheres=TRUE}}
}
\value{
No value.
}
\description{
Plot the given edges with \strong{rgl}.
}
\examples{
library(RCGAL)
library(rgl)
mesh <- Mesh(
  truncatedIcosahedron[["vertices"]],
  truncatedIcosahedron[["faces"]],
  triangulate = TRUE, normals = FALSE
)
# now we can plot the truncated icosahedron
tmesh <- tmesh3d(
  vertices = t(mesh[["vertices"]]),
  indices = t(mesh[["faces"]]),
  homogeneous = FALSE
)
open3d(windowRect = c(50, 50, 562, 562), zoom = 0.9)
shade3d(tmesh, color = "gold")
plotEdges(mesh[["vertices"]], mesh[["edges0"]], color = "navy")
}
