% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonReconstruction.R
\name{PoissonReconstruction}
\alias{PoissonReconstruction}
\title{Poisson surface reconstruction}
\usage{
PoissonReconstruction(points, normals = NULL)
}
\arguments{
\item{points}{numeric matrix which stores the points, one point per row}

\item{normals}{numeric matrix which stores the normals, one normal per row
(it must have the same size as the \code{points} matrix); if you don't
have normals, set \code{normals=NULL} (the default) and some normals will
be computed with the help of \code{\link[Rvcg]{vcgUpdateNormals}}}
}
\value{
A triangular mesh, of class \code{mesh3d} (ready for plotting
  with \strong{rgl}).
}
\description{
Poisson reconstruction of a surface, from a cloud of 3D points.
}
\examples{
library(RCGAL)
Psr_mesh <- PoissonReconstruction(SolidMobiusStrip)
library(rgl)
wire3d(Psr_mesh, color = "black")
}
