% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{qsqrt}
\alias{qsqrt}
\alias{print.qsqrt}
\title{Rational approximation of square roots}
\usage{
qsqrt(x, n)

\method{print}{qsqrt}(x, ...)
}
\arguments{
\item{x}{the positive integer whose square root is desired}

\item{n}{a positive integer, the higher the better approximation}

\item{...}{ignored}
}
\value{
A \strong{gmp} rational number (class \code{\link[gmp]{bigq}})
  approximating the square root of \code{x}.
}
\description{
Returns a rational approximation of the square root of
  an integer.
}
\examples{
library(RCGAL)
qsqrt(2, 7)
}
