% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Skull.R
\docType{data}
\name{Skull}
\alias{Skull}
\title{A points cloud of a skull}
\format{
A matrix with 383715 rows and 3 columns.
}
\usage{
Skull
}
\description{
A matrix containing some points forming a skull.
}
\keyword{datasets}
