% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-SpiderCage.R
\docType{data}
\name{SpiderCage}
\alias{SpiderCage}
\title{A points cloud of a "spider cage"}
\format{
A matrix with 10000 rows and 3 columns. The 10000 points have been
  sampled at random from a mesh containing more than 1700000 rows, obtained
  with help of the \strong{misc3d} package.
}
\usage{
SpiderCage
}
\description{
A matrix containing some points forming a surface looking like
  a spider and a cage.
}
\keyword{datasets}
