% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-StanfordDragon.R
\docType{data}
\name{StanfordDragon}
\alias{StanfordDragon}
\title{A points cloud of the Stanford dragon}
\format{
A matrix with 100250 rows and 3 columns.
}
\usage{
StanfordDragon
}
\description{
A matrix containing some points forming the Stanford dragon.
}
\keyword{datasets}
