% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-teapot.R
\docType{data}
\name{teapot}
\alias{teapot}
\title{A point cloud of the Utah teapot}
\format{
A matrix with 3644 rows and 3 columns.
}
\usage{
teapot
}
\description{
A matrix containing some points forming the Utah teapot.
}
\keyword{datasets}
