% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normals.R
\name{testinter}
\alias{testinter}
\title{Normals for a points could}
\usage{
testinter()
}
\value{
A function which takes just one argument: a numeric matrix with
  three columns, each row represents a point, and the function returns a
  matrix of the same size as the input matrix, whose each row gives one
  unit normal for each point.
}
\description{
Returns a function which computes some normals for a 3D points
  cloud.
}
\note{
The \code{getSomeNormals} function is intended to be used in the
  \code{\link{PoissonReconstruction}} function. If you want to use it for
  another purpose, be careful because the function it returns does not
  check the matrix it takes as argument.
}
\examples{
library(RCGAL)
testinter()
}
