% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tetrahedraCompound.R
\docType{data}
\name{tetrahedraCompound}
\alias{tetrahedraCompound}
\title{Compound of five tetrahedra}
\format{
A list with five elements, each one representing a tetrahedron
  by a list with two elements, the vertices and the faces.
}
\usage{
tetrahedraCompound
}
\description{
Five tetrahedra in a pretty configuration. Each tetrahedron
  is centered at the origin.
}
\keyword{datasets}
