% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ToroidalHelix.R
\docType{data}
\name{ToroidalHelix}
\alias{ToroidalHelix}
\title{A point cloud of a toroidal helix}
\format{
A matrix with 6000 rows and 3 columns.
}
\usage{
ToroidalHelix
}
\description{
A matrix containing some points forming a toroidal helix.
}
\keyword{datasets}
