% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convexhull.R
\name{convexhull}
\alias{convexhull}
\title{Convex hull}
\usage{
convexhull(points)
}
\arguments{
\item{points}{numeric matrix which stores the points, one point per row}
}
\value{
The convex hull.
}
\description{
Convex hull of a set of 2D or 3D points.
}
\examples{
library(RCGAL)
}
