% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{Mesh}
\alias{Mesh}
\title{Make a 3D mesh}
\usage{
Mesh(vertices, faces, triangulate = FALSE, merge = FALSE, normals = TRUE)
}
\arguments{
\item{vertices}{a numeric matrix with three columns}

\item{faces}{either an integer matrix (each row provides the vertex indices
of the corresponding face) or a list of integer vectors, each one
providing the vertex indices of the corresponding face}

\item{triangulate}{Boolean, whether to compute the normals}

\item{merge}{Boolean, whether to compute the normals}

\item{normals}{Boolean, whether to compute the normals}
}
\value{
A list giving the vertices, the edges, the faces of the mesh, and
  optionally the normals.
}
\description{
Make a 3D mesh from given vertices and faces; the returned
  faces are coherently oriented and normals are computed if desired.
}
\examples{
# a tetrahedron with ill-oriented faces:
vertices <- rbind(
  c(-1, -1, -1),
  c(1, 1, -1),
  c(1, -1, 1),
  c(-1, 1, 1)
)
faces <- rbind(
  c(1, 2, 3),
  c(3, 4, 2),
  c(4, 2, 1),
  c(4, 3, 1)
)
Mesh(vertices, faces, normals = FALSE)
}
