% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshes.R
\name{MeshesUnion}
\alias{MeshesUnion}
\title{Meshes union}
\usage{
MeshesUnion(meshes, clean = FALSE, normals = FALSE, numbersType = "double")
}
\arguments{
\item{meshes}{a list of \emph{triangular} meshes, each given as a list with
(at least) two fields: \code{vertices} and \code{faces}; the `vertices`
matrix must have the \code{bigq} class if \code{numberTypes="gmp"},
otherwise it must be numeric}

\item{clean}{Boolean, whether to clean the mesh (merging duplicated
vertices, duplicated faces, removed isolated vertices)}

\item{normals}{Boolean, whether to return the per-vertex normals of the
output mesh}

\item{numbersType}{the type of the numbers used in C++ for the
computations; must be one of \code{"double"}, \code{"lazyExact"}
(a type provided by CGAL for exact computations), or \code{"gmp"}
(exact computations with rational numbers); of course using
exact computations is slower but more accurate}
}
\value{
A triangular mesh given as a list with fields \code{vertices},
  \code{faces}, \code{edges}, \code{exteriorEdges}, \code{gmpvertices}
  if \code{numberTypes="gmp"}, and \code{normals} if \code{normals=TRUE}.
}
\description{
Computes the union of the given meshes.
}
\examples{
library(RCGAL)
library(rgl)

# mesh one: a cube; one has to triangulate it
cube1 <- cube3d() # (from the rgl package)
vertices <- t(cube1$vb[-4L, ])
faces <- t(cube1$ib)
mesh1 <- Mesh(vertices, faces, triangulate = TRUE, normals = FALSE)

# mesh two: another cube; one also has to triangulate it
cube2 <- translate3d( # (from the rgl package)
  cube3d(), 1, 1, 1
)
vertices <- t(cube2$vb[-4L, ])
faces <- t(cube2$ib)
mesh2 <- Mesh(vertices, faces, triangulate = TRUE, normals = FALSE)

# compute the union
umesh <- MeshesUnion(list(mesh1, mesh2))

# plot
rglumesh <- tmesh3d(
  vertices = t(umesh[["vertices"]]),
  indices = t(umesh[["faces"]]),
  homogeneous = FALSE
)
open3d(windowRect = c(50, 50, 562, 562))
shade3d(rglumesh, color = "red")
plotEdges(
  vertices = umesh[["vertices"]], edges = umesh[["exteriorEdges"]],
  edgesAsTubes = TRUE, verticesAsSpheres = TRUE
)
}
